Fixes from pull request
