+++ src/components/alert/index.jsx
@@ -1,7 +1,10 @@
 
 export const Alert = ({alertMessage, backgroundColor }) => {
-    const boxStyles = `block w-full text-white ${backgroundColor} text-center font-semibold p-4 mb-3`
     return (
-          <div className={boxStyles}>{alertMessage}</div>
+       <div
+          className={`block w-full text-white ${backgroundColor} text-center font-semibold p-4 mb-3`}
+       >
+          {alertMessage}
+       </div>
     )
 }
