+++ src/firebase/index.js
@@ -1,5 +1,5 @@
 import { initializeApp } from 'firebase/app'
-
+import { getAuth } from 'firebase/auth'
 
 const firebaseConfig = {
    apiKey: import.meta.env.VITE_FB_API_KEY,
@@ -11,5 +11,6 @@ const firebaseConfig = {
 }
 
 const app = initializeApp(firebaseConfig)
+const FirebaseAuth = getAuth(app)
 
-export { app }
+export { FirebaseAuth }
